Merge luiz_dev into master

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\RoomsController;
 use App\Http\Controllers\UsersController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -17,19 +18,19 @@ use Illuminate\Support\Facades\Route;
 
 Route::prefix('users')->controller(UsersController::class)->group(function(){
     Route::post('/', 'storeUser');
-    Route::post('/login');
-    Route::get('/{user_id}');
+    Route::post('/login', 'login');
+    Route::get('/{user_id}', 'show')->middleware('access_token');
 });
 
-Route::prefix('rooms')->controller('')->group(function(){
-    Route::post('/');
-    Route::delete('/{room_id}');
-    Route::post('/{room_id}/enter');
-    Route::post('/{room_id}/leave');
-    Route::delete('/{room_id}/users/{user_id}');
+Route::prefix('rooms')->middleware('access_token')->controller(RoomsController::class)->group(function(){
+    Route::post('/', 'createRoom');
+    Route::delete('/{room_id}', 'deleteRoom');
+    Route::post('/{room_id}/enter', 'enterRoom');
+    Route::post('/{room_id}/leave', 'leaveRoom');
+    Route::delete('/{room_id}/users/{user_id}', 'kickUser');
     Route::match(['post', 'get'], '/{room_id}/messages');
 });
 
-Route::prefix('messages')->controller('')->group(function(){
+Route::prefix('messages')->middleware('access_token')->controller('')->group(function(){
     Route::post('/direct/{receiver_id}');
 });
